Ensure that equality on unicode chars works.
Unicode characters were broken as it would only compare the first byte of
the character due to the strange way that string indexing works in
golang. This converts that to a working rune iteration which is far
safer and catches errors.

A regression test was added in order to test for this case in the
future.